"9:05am. I am up and my stamina is back up. Let me slack for a bit and then I will watch those two talks. Then comes review time.
9:55am. Almost ready to start.

https://www.youtube.com/watch?v=qRWTws_YPBA&feature=youtu.be
LDN Functionals #8 KC Sivaramakrishnan: OCaml multicore and programming with Reagents

Ok, let me go for it.

...For some reason in the latest version, the links feature is completely broken. I'll have to copy paste the URL into the browser.

10:25am. https://youtu.be/qRWTws_YPBA?t=1024

Hopac actually does have these features. I really need to test `<+>` to see if it does what I think it should do.

Actually let me give it a try now. I'll pause the video here for a bit.

```fs
job {
    let c1,c2 = Ch(), Ch()
    let! r = Promise.start (Ch.take c1 <+> Ch.take c2 >>- fun (a,b) -> a+b)
    do! Ch.give c1 1
    do! Ch.give c2 2
    return! r
    }
|> run
|> printfn "%i"

System.Console.ReadKey()
```

Let me give this a try. I'll try swapping the order. If this is really disjoint it should work.

```fs
job {
    let c1,c2 = Ch(), Ch()
    let! r = Promise.start (Ch.take c1 <&> Ch.take c2 >>- fun (a,b) -> a+b)
    do! Ch.give c2 2
    do! Ch.give c1 1
    return! r
    }
|> run
|> printfn "%i"

System.Console.ReadKey()
```

This deadlocks, but with `<+>` (and `<&>`) it works.

10:30am. Yeah, this is nice. Let me do one more thing...

```fs
let rnd_v = MVar(Random())
let wait t = MVar.take rnd_v ^=> fun rnd -> let t = rnd.Next(t) in MVar.fill rnd_v rnd >>=. timeOutMillis t
```

It is this thing again.

```fs
let rnd_v = MVar(Random())
let wait t = (rnd_v |> MVar.modifyFun (fun rnd -> rnd, rnd.Next(t))) ^=> timeOutMillis
```

Here is an alternative implementation that is more concise. `MVar`s can be used as a reference.

10:40am. The last primitive operator is `<&>` which Hopac has. I do not know how Hopac would compare performance wise, but all the ingredients are there.

10:45am. https://youtu.be/qRWTws_YPBA?t=1383

Ah, no, the `<*>` operator is not the same as here...I am not sure. Also can the `>>>` operator be implemented the same way as in here? I know that Hopac has those Kleisli operators, but I do not know whether or not they are atomic.

http://hopac.github.io/Hopac/Hopac.html#def:val%20Hopac.Infixes.%3E=%3E
> Creates a job that is the composition of the given two job constructors. (x2yJ >=> y2zJ) x is equivalent to x2yJ x >>= y2zJ and is much like the >> operator on ordinary functions.

Nope, it is not atomic.

11:05am. Did some Reddit posting and extended my original post with an extra on Reagents. I thought this video would be dull to those guys, but it was rather well received judging by the number of upvotes.

Now come the talk on Effects.

http://www.ccs.neu.edu/home/turon/thesis.pdf
Understanding And Expressing Scalable Concurrency

Inspired by the talk I posted, somebody also posted this. Looking at the TOC I see there is a section on backoff which I do not what it is at the moment, but as it is 299 pages I certainly won't be diving into it right now. This thesis would be an excellent place to start if I ever want to write my own concurrency library, which I do not ever expect to have to do.

11:20am. https://www.youtube.com/watch?v=0jI-AlWEwYI&feature=youtu.be
Alexis King - “Effects for Less” @ ZuriHac 2020

Let me take a look at this for a bit and then I'll be done with my study interlude.

11:40am. https://youtu.be/0jI-AlWEwYI?t=975

This is interesting. Well, having things in multiple modules won't be any problem whatsoever in Spiral.

11:55am. https://youtu.be/0jI-AlWEwYI?t=1704
> The simplest, most naive solution that does not work is type dispatch.

This has my interest considering I am going to use it in Spiral.

https://youtu.be/0jI-AlWEwYI?t=2910
> So there is various reasons why I think whole program specialization is really not viable and I think pursuing this is a lost cause and it is hopeless, and we really should not try to commit to whole program specialization.

And yet it is the only way to cleanly combine different languages. Hard to do a Cuda backend with the kind of strategy GHC uses. It would take some work, but it would be possible to combine whole program specialization with JITing.

https://youtu.be/0jI-AlWEwYI?t=2950
> Effect systems are fundamentally about dynamic dispatch.

I have no opinion on whether algebraic effects are fundamentally about this, but I would not assume that it follows based on the previous premise.

1:05pm. Done with the talk. The thing is, the issue with bind being slow at runtime is also why it would be slow at compile time in Spiral. I've thought a bit about inlining during the prepass, but could not get far with that.

Hmmm...

Well, things like state and reader won't be used as often as they are in Haskell so forget that.

1:10pm. I think I agree with her that algebraic effects need dynamic dispatch if they are not monadic in nature.

...Hmmm, could I implement something like compile time delimited continuations? It would be possible to just grab the program slice at compile time without necessarily having to do filtering. That would make monads a lot faster to compile in Spiral.

I'll leave that as an future exercise. It probably won't be necessary so I won't worry about it.

1:15pm. Let me have breakfast here."