"2:20pm. Done with breakfast. Let me read Scarlet's novel for a while. There is not much out so I should catch up quickly.
https://mangadex.org/title/38088/may-i-please-ask-you-just-one-last-thing
Alternative title: The B*tch Who Punches Everybody

I found this a few days ago and it is amazing. The bitch is undeserved - she is a classy lady.

3:05pm. Kengan Omega and then I start.

I need to do at least a little today. Just the review is not enough. I'll be laid up over the weekend so I want to conserve some time.

3:15pm. Done slacking. Let me focus now.

Spiral v0.2 starts here.

4pm. Er, I had to take a 45m break due to a distraction.

4:10pm. It is no use, I am completely distracted.

4:20pm. This is not going to work. In the past two days, my mind has been entirely on the review. Even though it is the time to start, I hadn't spent any time at all charging myself for the task at hand.

4:25pm. I think I'll step away from the screen for a while. Since my thoughts are distrupted to this degree, I need to find my center.

The way things are going, I am not going to be doing any programming today. Instead I should just focus and crystalize the first step in my mind.

To be honest, I am not sure if I want to just plugin the current version of the language in to VS Code as I planned, or whether I want to start the redisign right away. I should settle that mentally.

On paper it does not seem like it is anything important, but trying to do a thing without fully commiting to it is a mistake."