"9:45am. I am up. Thanks to yesterday, I got rid of some of my flatfottedness. Damn, was that thing annoying.
What is important in life? One's own power.

Never forget that. As a matter of principle, debts are to be cut. Today, I will finally be able to focus on programming for a while.

My mind feels a bit drained though. I haven't slept too well.

Still, when has that ever stopped me?

10am. Let me waste time for a bit first though as I am too groggy. I haven't really had the mind to think about the task that I actually need doing.

10:40am. Let me finally start. My will is starting to go up.

10:45am.

```fs
                values =
                    Map.empty
                    |> Map.add "infinityf64" (Lit (LitFloat64 infinity))
                    |> Map.add "infinityf32" (Lit (LitFloat32 infinityf))
```

Ah, I forgot about this. Yeah, I am going to allow top level values just to eliminate this dependency.

```fs
            | Some(Forall _) -> raise <| CompilationError "main has to be a regular function, not a forall."
            | Some _ -> raise <| CompilationError "main has to be a function."
            | None -> raise <| CompilationError "main has to be present in the last module."
        | _ -> failwith "Compiler error: Module that has been compiled should always be in the environment."
    with
        | :? CompilationError as x -> Error x.Data0
        | :? TypeError as x -> Error(show_trace settings x.Data0 x.Data1)
        | :? CodegenError as x -> Error(x.Data0)
        | :? CodegenErrorWithPos as x -> Error(show_trace {settings with filter_list=[]} x.Data0 x.Data1)
```

I can't just take this function and call it directly as the positions get stripped.

10:50am. Ok, how about for starters I split the trace and the final error message?

10:55am. Let me start it like that. The goal is starting to form in my mind.

The way the language is implemented now, I can't do much with it. At most I will be able to do some primitive compilation. I can't even use it as a caching server.

So my goal for the mean term will be modest - let me jack it into the server. Just that will serve as the appetizer.

After that I should just start on the redesign starting with parsing.

Right now the Spiral v0.2 parser's error messages are broken due to getting eaten and this is something I should address at the onset.

Also I really should break out the disassembler and take a look at how well F# is inlining various expressions and data structures because right now that is just awful. I need to optimize this.

11:05am. Since I will have editor support, the parser parts I should test in advance of everything else. I should only move to typechecking once that is done. Even though I won't be able to run the programs I am making I should leave the prepass and partial evaluation for last.

11:10am. There is just one last thing...

Foralls. I'd like them to be unicode, so that I could write \a and have the editor convert them to ∀. That would be much better than having to write forall explicitly. This is something I should on the VS Code plugin page. Come to think of it, I still have not gotten an answer to my question.

https://marketplace.visualstudio.com/items?itemName=brunnerh.insert-unicode

Hmmm, I should look into how this plugin does it.

11:15am. Since just jacking the `compile` function into the server won't be much of a challenge, what I really need to focus is gather my will for the parser and the typechecker parts. I need to focus on the redesign itself.

Let me take a little break here.

11:40am. Let me have breakfast here. No much point in starting anything new now.

Thought I am still not fully focused, I did get a feel for wanting to do the parser just for a bit now.

12:40pm. Done with breakfast. Let me resume.

Honestly, I feel like stepping away from the screen for a while. Whereas yesterday I got distrupted, right now my mind feels too much on edge. I feel both intense heat, and anxiety and tedium.

I want to step away from this for a while. I want to work on this, but not under such high mental strain. I've sharpened my mind, but a sharp mind needs to be cool.

I'll step away from the screen and the computer. It is good to do that sometimes."