"8:55am. Right now I am still recovering from that 10/10 trolling attempt on my life. Yesterday and the day before I was incandescent the whole time and laid out in bed. Right now I am in a more normal state of mind, but feel like resting for a while. Since I've worked hard for so long, I deserve an vacation every now and then. I'll resume programming when I feel like it.
If there is one message I am going to take from this event, it is that it is time for me to start developing my mask of professionalism. I'll need it anyway for important things down the road once I am done with Spiral and looking to get my hands on those chips.

I even at this age retain too much of my juvenile high school mentality. Even though the school is long over, I am still in school and this is a good time to graduate. Schools are such an abnormal, forced situations. When you are an adult, you have the freedom of association, but in school you are forced to endure the tedium and associate with those you dislike. I did not get much for being an adult, but from here on out I am going try and play social situations more like DCSS as opposed to Diablo.

I am going to get both more dignity and free up my schedule a little as a result. It is a win-win.

9:15am. Unlike usually, I won't aim to do any programming in the morning today. Like in January I think I'll try to vacation a bit. The best way to start work isn't to force yourself, but to just get tired of the triviality of wasting time."