"5:50pm. Let me read chapter two of Dungeon Meshi and then I'll just do a tiny bit of programming.
6:05pm. I went from...

```fs
    with
        | :? CompilationError as x -> Error x.Data0
        | :? TypeError as x -> Error(show_trace settings x.Data0 x.Data1)
        | :? CodegenError as x -> Error(x.Data0)
        | :? CodegenErrorWithPos as x -> Error(show_trace {settings with filter_list=[]} x.Data0 x.Data1)
    |> fun x -> timings.Elapsed, x
```

...to...

```fs
    with
        | :? CompilationError as x -> Error([], x.Data0)
        | :? TypeError as x -> Error(x.Data0, x.Data1)
        | :? CodegenError as x -> Error([], x.Data0)
        | :? CodegenErrorWithPos as x -> Error(x.Data0, x.Data1)
    |> fun x -> timings.Elapsed, x

let compile settings x =
    let tim, er = compile' settings x
    tim, er |> Result.mapError (show_trace settings)
```

This is the compile function.

I just separated it into two.

Now in the server I will be able to reuse `compile'`.

```fs
type SpiralModule =
    {
    name: string
    prerequisites : SpiralModule list
    description : string
    code : string
    }
```

Ok...note that this is not an exactly flattened project structure in the vein of F#. It is linear, but it is more like every file is its own package even though they are called modules. It was conventient to do it like this when I was writing Spiral modules directly in F#, but I'll have to change this.

6:20pm. This is just an apetizer for tomorrow.

```fs
        | File x ->
            failwith "Not implemented yet."
```

Tomorrow what I will do is start work on this segment. I should also do a command to trigger compilation. Spiral in its current form is too slow to do it from scratch on every keypress. It just won't be viable until I start caching intermediates, and even then the full pass which does prepass + partial evaluation + codegen should only be on command.

6:30pm. I'll have to rewrite the language from scratch again, but just doing this first step of integration with VS Code won't be too difficult.

Right now I am ass-blasted, but I will get back into it before long. This whole thing will be good in the long term. By weeding out the bad, the good can grow.

I hope that in the future I will be able to nurture some connections in professional capacity. As things are now I am not really going to have my goals align with anybody around me. At this point I feel like an outsider if not in my own family, then definitely in the area I live in. The sooner I finish the language and get those chips the better. Even though it would not be money on its own, just finishing that ensemble poker agent would be an immense burden off my chest. I've always pinned my hopes on RL, but I was not prepared for the harshness of the path. I need to take the hammer and bash the pin all the way in. I just want to win in the way I desire.

In Spiral v0.09 doing the ML library was actually particularly satisfying. If only I could actually make it useful that would be something. I need a better finisher this time.

I admit, I thought that Deepmind or OpenAI would come up with something useful by 2020, but those guys are way more useless than I thought. They are overpaid clowns with good PR departments. I did not think I would have to reach for the next batch of hardware just to get a basic agent going to a complete satisfaction."