"8:40am. Last few days are really unfortunate. Until my obsession is through with a subject, I can't really tell it to stop. It just so happens it was not directed at programming tasks in that time. I am still asspained, but not so much that I cannot think. I should be able to get some programming done today. Tomorrow I'll proofread the review. I think I'll just post the second part which is the transhumanist conception of power. I doubt anybody would care about the trivia of what I was doing in the last six months.
I need to rewrite it just a tad to highlight its evolutionary as opposed to revolutionary nature.

8:50am. Let me chill for a while.

It is true in all things - you can only control yourself to a limited degree, so it makes more sense to focus those efforts on controlling the environment around you. It is not just in the social regime - me creating Spiral is an act of determination that I am not going to accept Python or consent to adjusting my style so it is lesser.

The way I am acting here is the way I should be acting everywhere; I won't resent others for my low status, instead I should see it as my punishment for acting in an undignified manner. Eventually, I am going to be done with Spiral to a satisfactory degree and I will be able to do what I want with it. In the case of emergency where I **need** cash ASAP, I do have my F# and .NET skills to fall back on.

I am really close to being a full spectrum programmer right now. Once I finish v0.2, I'll have leveled up once more. I just need more concurrency knowledge. Whether Rx or Hopac, in a realistic application I would be using those all the time. I still need more experience with them.

9:05am. It is kind of like this - I never want to get a job, but it would sure annoy me to find out that I could not do it. Regardless of whether I'd manage to get through the interviews (I highly suspect I would, but it is not something I can assume) I intend to make it so that any kind of task that would be dropped in front of me is doable and is so without code monkeying.

I want to so that every line I ever written when it is studied by future historians is seen as - 'Wow, this guy sure was good compared to those other monkeys.'

9:10am. Given the kind of life I led of just letting myself be carried by the flow, maybe in the depth of my soul I always had an yearning for this kind of control. Spiral is my outlet for that. Maybe that I need Spiral to cause the Singularity is just an excuse.

My reasons for it are good - a fully fledged Spiral would be a great asset, but the lazy approach of just using the language of the time for everything should be viable. The ML script kiddies are not wrong for picking Python. Saying that Spiral is necessary would be going too far in the present time.

That having said - I don't know yet, neurochips might change that equation. So I'll cover that eventuality. I've said as much 5 years ago.

If possible I'll do this kind of thing for both myself and as a service to the community. Making weapons is a noble endeavor - they make it easier for both myself and others to hunt.

If it turns out that Python is the be all of languages, and Deepmind and OpenAI reap all the benefits of the next ML wave as well, then (maybe) I'll accept that success was just not meant to me in this life. I'll console myself that I managed to attain a modicum of skill even though not much else.

Either way I need to remember the pride of somebody living in this era.

From the start to the end, computers are the most miraculous invention of humankind. And games are the highest form of art. People live and die for all kinds of things. Even if I never get my power, and end up being skull and bones on some dungeon floor - a warning to adventurers, that is fine.

This should be worth trying in at least one life. Humans are much stronger when they are in pursuit.

9:35am. Based on how my thoughts are flowing, I will get back into it before long.

9:40am. Let me do my morning reading. I think that I won't push myself to program today either.

I know that it seems like waste of time, but distruptions do happen, and vacations are necessary from time to time. This is a great time to heal myself a little from all the stresses of the past six months.

Since this is me, I doubt this will last for too long."