"9:20am. Yesterday I've really gotten into Dungeon Meshi and am halfway through. It is an amazing ride. Now that I do not have work, I can really get into it properly. 10/10. Had I read something like this 1-2 decades ago, it would have been my reason for getting into manga. It was the SnoopyCool and Ripper's Anime era when I started this trip.
9:25am. How about I continue my vacation into today?

If I look inside myself, it is not like I lack the determination necessarily. I do feel it, but I still feel overloaded. I should just read Dungeon Meshi for another day and when I am over this issue dive in. A lot of stuff needs to be done to do editor support, but I can only focus on programming when I am not worried about other things. It is not like I need to do it literally ASAP.

It is important, but not so much that I have to act like a robot about it.

9:35am. Let me just play around today as well. Proving my determination is important, but understanding that I am free is important as well.

Being able to play games or read manga without any other concerns is an amazing feeling that I lost when I started programming.

If you start programming in the morning and finish at 6pm, even with breaks inbetween, it is not the same thing as having the whole day to yourself. Even if I stop at 6pm on paper, I am mentally exhausted at that point and can't really enjoy things properly. It is also hard for me to plug out of the programming mindset.

9:40am. I'll indulge myself for just a little bit longer. I'll let Dungeon Meshi heal me and then I will start.

I should post that review today on the Simulacrum blog, but I do not feel like it. Forget that for now. I'll leave work for later. Today I do not even want to touch upon that."