"12:55pm. https://cybertrophic.wordpress.com/2019/09/30/why-posthumanism/ > Those major tenets being (1) the existence of some kind of essence of human-ness, a universal and eternal human nature that inheres in all of us and makes us fundamentally “the same” in a way that has profound moral, social and political implications, and (2) the status of humans as individuals, as indivisible selves who are the authors of their own thoughts and actions, and who have free will and agency. I believe that both of these tenets are brought into serious question by modern science and philosophy, and that they should at least provisionally be rejected.
This blog is interesting.

> And their progress has had decidedly posthumanist implications, as when Peter Singer comes to the conclusion that the pleasures and pains of all animals must count equally, not just those of humans.

Though as soon as you conclude that, you also then follow it up by concludding that evil exists for a reason.

Once you stop seeing evil as an unfortunate accident brought on by Satan but a practical necessity, you can start to make progress towards viable system of morality. The humanist tenents are a good thought experiment, but taken seriously amount to nothing more than a lobotomy.

> For a child, growing up means letting go of solipsism, letting go of the idea that one is the centre of the universe.

Nobody really believes in solipsism. Reality has a way of discouraging that practice. I haven't really been convinced that this is poor as a goal. I mean sure, seeing yourself as a product is factually true, but where can you push with that?

If you can create something as absurd as the self - which you need in order to assign ownership to power, then you can see some other ideas in a new light. Letting go of solipsism is not a sign of growing up. You can't let go something you've never had. You can only let go of the idea of having it. Not dwelling in daydreams is growing up, but too many people, lessers all of them, decide that this means the dream is not worth it.

> For (post)humanity, escaping from our infancy means engaging in the same kind word of Copernican turn: no longer seeing ourselves as something unique, special and indivisible, no longer seeing ourselves as magic, godlike minds wearing bodies, but rather as storms of flesh, continually produced and reproduced in a universe of material systems.

If you are going to go to war, what is first needed is a dream to draw the attention away from the harsh reality of combat. The person who only looks at what is in front of him is weak.

If one is going to be realistic, one needs to be realistic about ideals themselves.

1:20pm. I sure don't feel like fighting now, so I won't post this on his blog.

Usually I count my day as lasting till 6pm, so I always feel pressured, but now I am assuming I have till 12am to rest so it feels like I have a lot more free time. Tomorrow, I will endeavor to do some programming. Not because I have to, but because I want to. Because it is the right thing to do. Not because it is my job, but because it is my way of life.

After I integrate the mental changes I am making, I won't be split anymore. I made mistake being 50% programmer, and 50% high schooler (or before that, 50% trader/50% high schooler) for a long time now. I will simplify my life and strenghten my core by doing that. For somebody who is 98% of the time in his room, my life is too complicated and in need of cutting.

I'll find what I want through programming and nothing else."