"8:55am. I am up. And given that my thinking has gone back way to middle school, I am almost done resting and integrating my mental changes. Though I say changes, I am mostly removing the overly complicated mental scaffolding that lead me to figuring out the self improvement loop. The way I went at it was difficult and I am tracing my thinking all the way to those bygone days decades ago. Maybe it made sense to act the way I did at the time, but now that I have the insight of how superpowers can be attained through the use of computers, all those ways of thought seem like a waste. It is just dead machinery that I never had the courage to turn off. I have so many regrets and they constantly assault me as daytime nightmares because I can never accept them.
I find that when I no longer have to justify the past, I can just look back and say: 'I wish I had acted in a different manner.' Even if I would have lost the spark by doing that. My desire to say that out loud is stronger than anything else.

I am treasonous to my core after all. I care not for outdated traditions, apparently not even my own. Though the trolling attempt was the trigger for this, no doubt my subconscious would have found some other excuse given that going forward into the future I will need to be status sensitive. I'll only step one foot into a professional setting, but I do not want it to get chopped off.

I am so damn cheap - now that I have the main principles of transhumanism, the sacrifices my past self made to get to them seem to matter not at all. Celebrating my past heroism is too expensive.

But well, this kind of nature is what makes me great as fucked up as it is. From the bottom of my heart, I love this part of me. Without this kind of decisiveness I would be hopeless. From a rational perspective, I do feel it is actually better to be treasonous than respectful here.

9:10am. When the Singularity happens, the new gods will be chosen. I can only be grateful it won't come down to a vote.

I can't deny it - when the building is done being constructed, it is time to put away the equipment as useful as it was. I became too attatched to it.

...

You know what, since I am still thinking about it and putting my mind in order, let me extend the vacation. It should not be a matter of pride that one works all the time.

I can afford doing it for a full week or two without being branded a slouch. Since the trip through the memory lane is nearly over, I expect that I'll get back to it soon, probably tomorrow.

One thing I will definitely do today is post that essay on the Simulacrum blog.

9:35am. Do I feel like doing it now? I'll do it later. Let me do my morning reading."