"9:25am. I am up. Let me do my morning chilling. I did the right thing by taking the time off to vacation, I really do feel that the thoughts I have made during this period will be crucial towards getting my life in order. They'll only affect a tiny part of my life, but it is dangerous to have chinks in the armor. I need to be an active player in all areas. I do not want to have any weaknesses.
10:25am. Let me start. Unlike in the past week I neither feel stressed nor distracted. Likewise, neither do I feel like not working on this. So till the day is over forget the distractors and focus on what is really important.

Programming - it all comes down to it in the end.

I can only prove my worth by finishing Spiral.

10:30am. It is finally the time to get back to my usual pace. Today feels quite regular.

https://code.visualstudio.com/api/language-extensions/syntax-highlight-guide

Let me start with this.

10:40am. I lied a bit. I am still slightly distracted, but this is around the usual for me. Still, going forward I want to exceed the level of focus that I had before. I'll let this develop naturally. Eventually I will get tired of the past story and start getting interested in the new one. I won't rush this thing.

https://code.visualstudio.com/api/language-extensions/syntax-highlight-guide#scope-inspector

I never tried this out before. Though before I do, let me read the article.

> Starting with release 1.43, VS Code also allows extensions to provide tokenization through a Semantic Token Provider. Semantic providers are typically implemented by language servers that have a deeper understanding of the source file and can resolve symbols in the context of the project. For example, a constant variable name can be rendered using constant highlighting throughout the project, not just at the place of its declaration.

I really should take advantage of this.

11am.

```json
{
  "scopeName": "source.abc",
  "patterns": [{ "include": "#expression" }],
  "repository": {
    "expression": {
      "patterns": [{ "include": "#letter" }, { "include": "#paren-expression" }]
    },
    "letter": {
      "match": "a|b|c",
      "name": "keyword.letter"
    },
    "paren-expression": {
      "begin": "\\(",
      "end": "\\)",
      "beginCaptures": {
        "0": { "name": "punctuation.paren.open" }
      },
      "endCaptures": {
        "0": { "name": "punctuation.paren.close" }
      },
      "name": "expression.group",
      "patterns": [{ "include": "#expression" }]
    }
  }
}
```

Now that I am really studying this, I see that this is actually a mutually recursive definition.

11:05am. Focus me, stop taking a peek if you've gotten a reply to that parser combinator thread.

11:10am. `\l` is really easy to write. If I am going to use Unicode in order to get the forall, I might as well introduce the Unicode lambda.

11:20am. Let me take a short break.

11:50am. I am back. Let me resume reading the guide.

It is coming to me. Though I'll use the semantic token provider, I should implement basic keyword highlighting.

12pm.

> There is one additional complication for injection languages embedded languages: by default, VS Code treats all tokens within a string as string contents and all tokens with a comment as token content.

Definitely going to need to think about this...

12:10pm. Done studying the document. I wish I could say I am focusing my whole being on this, but it feels like I am giving it 20-30% of my all at best. Well, it is fine. Reading documentation was never exactly a great locus for my motivation.

At this point, I am mostly daydreaming rather than working. Still, I think I get most of the things in the document, what I do have to do next is get familiar with...

https://macromates.com/manual/en/language_grammars
https://macromates.com/manual/en/regular_expressions
https://www.apeth.com/nonblog/stories/textmatebundle.html

12:25pm. Ok, I've decided. Forget this. I am going to go with the server side tokenization, not because I do not want to use regular expressions, but simply because I do not want to duplicate effort. If I start using Textmate grammars, then I will be putting stuff on the client side which I do not want. I should strive to minimize the amount of code I write on the Typescript side.

12:30pm. Now that I've confessed to this the goal I should take on is clear - do the tokenizer for Spiral. I already have one, but the one I need needs to be incremental and fit for use in the server.

I admit, I haven't though that I would be doing highlighting using the tokenizer, but that is great as it gives me a natural thing to strive for. It seems a bit convoluted to go through the server, but as syntax highlighting is the very first step, this is a great opportunity to do some ground up work.

12:35pm. If only I could get the rest of my mind to look straight that would be great.

Once I start programming, my focus should go up. And that is what I am going to do after the break.

In the tokenizer, I am going to get rid of the dependency on FParsec so I can detect tabs properly. At this point I do not actually understand how to do deal with changes just yet. So before I start, that is the first thing I should check out. Can I get the initial file as one big change? That would be great. If not, I'll just do a bit caching in the plugin to make sure that the server gets the proper info.

...I'll know if I try to save a file with no changes whether VS Code is paying attention to this kind of thing. Probably not...

12:45pm. I'll leave it for later. Let me get some food."