"3:45pm. No, the old prototype is no good. I'd be better off doing it from the top. The main focus of it was to do the typechecker, but here I think I will just ommit that part for the time being and make parsing the latest stage with a few others inbetween. Like tokenization, blocking and such.
Just the blocks make things a lot more difficult than before.

3:50pm. Figuring out how to do deal with them is my number one challenge right now. Once I get some inspiration regarding them, I will be able to deal with everything else. I can think of inefficient, but easy ways of dealing with them, but I can't think of how to get both efficiency and easy. I tend to equate ease with elegance.

Elegance is not an intrinsic property of the code, but rather it is a subjective measure of how easy the code is to store in one's mind.

3:55pm. Ok, me. Stop daydreaming so much and start writing code.

I still have a bit more left in me. I should strive to do this for an hour more.

4pm. Ok, ok, so what should be the first thing on my list?

How about a file server? Not even something that handles blocks, but just an array of lines?

4:15pm. Now I find myself surfing /g/. No forget this. Let me just step away from the screen for a while. I'll turn off the computer. I am just so completely dry as to what I want to build here. Even the simplest thing takes me a lot of charging before I can start doing it.

Let me stop for an hour or so. After that I'll call it a day."