"6pm. I spent the last two hours away from the screen just thinking. And I've figured it out. The whole pipeline!
In that sketch that I did today, the insert and remove functions iterate from the start on every change. This is something I am going to do in the pipeline as well. There is just no way to get around that as far as I can see. But thankfully that should not be a serious overhead.

I will be able to use string references and weak tables in order to do memoization - alternatively, I might stick a reference cell. There are all sorts of ways of doing this.

The way I will get around the issues of parsing comments is to split blockization into parsing the comments and then the statements. Then I will separate out the comments from the structure. Multiblock structures I will do in a separate step. I actually do not need multi block structures to do parsing, only the typechecking and the later steps need it.

Even though the whole file will be refreshed on every change the whole thing will be fast because I will keep track of the latest changes. Then I'll only have to redo the work from the dirty block - getting to it will be fast. And the blocks past the dirty ones will be recalculated during the normal course of processing.

6:15pm. I am actually pretty satisfied right now. Tomorrow I should be able to do some really effectful programming. The system I have in mind is actually simpler than the sketch I wrote today.

Right now, let me do an experiment. Rather than using weak conditional tables, it might be better to use a second array...

Ah, actually nevermind. I would be using a resizable array in that case.

```fs
let r = Array.zeroCreate 10
match r.[0] with
| None -> printfn "None"
| Some i -> printfn "%i" i
```

Ah, so None does get compiled to nulls. As I expected. Don't really need it though.

6:20pm. I close for the day here."