"9:15am. I am up. Let me chill just a bit and then I will start. I do not want to dawdle too much here.
9:30am. Let me start.

9:35am. I am thinking. The feedback I am getting right now is good. Definitely.

9:45am. Hrrrmmmm...

Let me talk about the prototype. In the prototype I separated the parsing and the typechecking, but it would be easier to make parsing more like typechecking - one file at the time.

One issue with blocks is that I do not know to do the goto for the typechecker. In the prototype, the supervisor was responsible for that, but it would have been easier if it just reacted to what the parser sent it. No muss, no fuss.

It is true - I decided on that architecture in order to allow parsing and tokenization of files that aren't in the project itself, but that is not that important. It would be better not to waste CPU cycles instead.

The most important feature of editor support is that I give the user feedback as fast as possible. Everything else is secondary to that. Everything else must be subordinate to that.

10:10am. You know what, I am going to give up on the compile in the background feature. It works for F# because its compilation cannot display further errors, but Spiral can. Both the prepass and the partial evaluator will be able to return type errors. So I will have doing actions while the compilation is going on interrupt it.

10:15am. Let me take a little break here.

10:35am. ...

10:45am. Right now I am wracking my brain how to minimize the role of the supervisor.

10:55am. I said I figured out the pipeline, but that is true for just one. I can think of 3 different pipelines...

1) Parsing only - for files that are not in the project.
2) Parsing + typechecking - for project files
3) Parsing + typechecking + prepass - for compilation

I need the ability to switch between them. Plus when the third thing is done, I need the ability to switch to partial evaluation + codegen.

11am. I can't really figure out how to do this neatly using streams.

...Hmmmm, I think I have it.

I thought of this idea in the context of Rx, but I can do it with Hopac as well.

I can take...

```
3) Parsing + typechecking + prepass
```

And turn it into...

```
3) Parsing + delay P + typechecking + delay T + prepass
```

What the two delay nodes would be doing is acting as modulators. They'd have two modes - blocking and forwarding.

In blocking mode the delay node would just buffer the inputs without sending them on.

In the forwarding mode, the delay node would pass all the inputs forward in a synchronous fashion. When switching modes from blocking to forwarding, the node would first forward all the old inputs before doing any receives.

11:10am. Hmmm, come to think of it I am going to need this kind of thing even without switching modes.

```
1) Parsing only - for files that are not in the project.
2) Parsing + typechecking - for project files
```

When moving attention to the previous file in the order, I'd want to switch to parsing only temporarily.

11:15am. But when switching back it would not be enough to just resume - I'd need to replay the past inputs from the start for the typechecker. I guess I'll just have the two delay nodes differ in their switching behavior.

11:25am. Hmmm, maybe I should have the result of a compilation of a file be an MVar.

...What I could do is have the MVar act as a release trigger for the typechecker delay node.

11:35am. The same for the prepass. And I could use MVars like that to start compilation as well.

11:40am. For a bit now, it feels less like regular programming and more like circuit design that I am doing here.

https://www.youtube.com/watch?v=bJCEN2-uyjY&list=PLfbK4pt7rdhjsMQQ_E5jVe98cNhlrdcle&index=189
Plutonia Revisited Music - Map 27 - Planned Overload (Good Doom Music #91)

Plutonia Revisited really has excellent music all around.

11:50am. What I am doing this morning is really up to my ideal standard. I might not have written any lines of code, but I've been deeply engrossed in the task at hand.

...No, I am not going to use an MVar as that would make the pipeline async. Rather, I will have the pipeline reset itself. I'll use a channel rather than an MVar. When I write None, that would mean to go into blocking mode, and Some would mean to go into forwarding mode.

12pm. https://ollef.github.io/blog/posts/query-based-compilers.html

I've read this. Let me check out that library he is talking about just for a bit.

12:10pm. I have no idea what to think about this. There is too little in the way of example for that Rock library, and it is all in Haskell to boot.

Yeah, it would be easiest if memoization was the silver bullet for compiler construction, but memoization is poor at responding to changes. It is add-only kind of thing.

https://www.microsoft.com/en-us/research/uploads/prod/2018/03/build-systems-final.pdf
Build Systems à la Carte

12:20pm. I have no choice, I am going to have to look into this.

A central cache that can manage invalidation is interesting, but I do not know how to do it.

As is expected there are all sorts of options here.

Let me take a break here."